requirements.txt
discord.py
praw
openai
pytz
requests
